app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
asyncpg==0.26.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.68.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.0.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic==1.10.12
pydantic_core==2.6.3
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing_extensions==4.8.0
uvicorn==0.15.0
python-dotenv